+++ Acceso a datos/src/main/java/com/example/gym/service/ReservaService.java
@@ -1,6 +1,5 @@
 package com.example.gym.service;
 
-import com.example.gym.domain.Cliente;
 import com.example.gym.domain.Reserva;
 
 import java.util.Optional;
@@ -10,6 +9,6 @@ public interface ReservaService {
     Set<Reserva> findAll();
     Optional<Reserva> findById(int id);
     Reserva addReserva(Reserva reserva);
-    Reserva modifyReserva(int id, Reserva newReserva);
+    Reserva modifyReserva(long id, Reserva newReserva);
     void deleteReserva(int id);
 }
